using bubble sort

+++ 0-bubble_sort.c
@@ -8,7 +8,7 @@
  */
 void bubble_sort(int *array, size_t size)
 {
-	long unsigned int i, j;
+	size_t i, j;
 	int temp;
 
 	for (i = 0; i < size; i++)
